+++ src/classes.js
@@ -46,10 +46,11 @@ $.fn.extend({
         if (not(cls) || (""+cls).trim() === "") return this;
         return this.each(function(){
             var el = this;
+            var hasClassList = typeof el.classList !== "undefined";
             $.each(cls.split(" ").filter(function(v){
                 return (""+v).trim() !== "";
             }), function(){
-                if (el.classList) el.classList[method](this);
+                if (hasClassList) el.classList[method](this);
             });
         });
     }