+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 + [x] Events: add `$(document).ready(...)`
 + [x] Events: fix context for event handler function
 + [x] Init: add support for array. Now you can compose `$` from array of `$`
++ [x] Events: improve method `on`
++ [x] Promise: add method `always(onAlways)`
 
 ### 1.0.2
 + [x] Events: fix `function.name` property for IE11